feat(bex): move components into hub/src/app/pools/components
