+++ apps/hub/src/app/pools/components/create-pool-initial-liquidity-input.tsx
