+++ apps/hub/src/app/pools/components/ownership-input.tsx
